requirements.txt
APScheduler==3.11.0
blinker==1.9.0
click==8.1.8
dnspython==2.7.0
email_validator==2.2.0
Flask==3.1.0
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mysqlclient==2.2.7
python-dotenv==1.0.1
python-http-client==3.3.7
sendgrid==6.11.0
SQLAlchemy==2.0.39
starkbank-ecdsa==2.2.0
typing_extensions==4.12.2
tzlocal==5.3.1
Werkzeug==3.1.3
WTForms==3.2.1
